Changed GenericLiteralVisitor to treat float literals as Double by default (#3074)
Currently, when a floating-point number is provided without a suffix in Gremlin Lang (e.g., 1.0 or -123.456), it's automatically converted to a BigDecimal. This differs from most programming
languages where floating-point literals are typically treated as double-precision values.

To align the behavior of Gremlin Lang with common programming language conventions and user expectations, this changeset changes GremlinLangScriptEngine via GenericLiteralVisitor to treat float literals without explicit type suffixes (like 'm', 'f', or 'd') as Double by default. Users who need BigDecimal precision can still use the 'm' suffix (e.g., 1.0m). Gremlin Groovy will still default to BigDecimal for float literals.

+++ gremlin-core/src/main/java/org/apache/tinkerpop/gremlin/language/grammar/GenericLiteralVisitor.java
@@ -476,7 +476,7 @@ public class GenericLiteralVisitor extends DefaultGremlinBaseVisitor<Object> {
             // parse D/d suffix as Double
             return new Double(floatLiteral);
         } else {
-            return new BigDecimal(floatLiteral);
+            return new Double(floatLiteral);
         }
     }
